Wow what a day
This time it's some other plugin messing with the `i` again.
So... When typing the previous commit I've noticed that when I try to
type `time` right after I type `ti` it becomes `t1i`. The real reason
and conditions are unknown to me. But `silent = true` in the
keymap settings "magically" works. Should read on that later...